requirements.txt
aws-cdk-lib==2.29